pyproject.toml
[tool.poetry]
name = "foundry-instance-manager"
version = "0.1.0"
description = "A command-line tool for managing Foundry VTT instances"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "foundry_manager"}]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
pyyaml = "^6.0.1"
requests = "^2.31.0"
jinja2 = ">=2.10,<3.0"
markupsafe = "<2.1"
rich = "^10.0.0"
docker = "^7.1.0"
Pillow = "^10.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pre-commit = "^3.5.0"
build = "^1.0.3"
setuptools = "^69.2.0"
git-changelog = "^0.4.0"
types-requests = "^2.32.4.20250611"
types-docker = "^7.1.0.20250523"
types-setuptools = "^80.9.0.20250529"
types-pyyaml = "^6.0.12.20250516"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.poetry.scripts]
fim = "foundry_manager.cli:cli"
